requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.13.0=py37haa95532_0
aiohttp=3.7.4=py37h2bbff1b_1
anyio=3.2.1=py37h03978a9_0
argon2-cffi=20.1.0=py37hcc03f2d_2
astor=0.8.1=py37haa95532_0
async-timeout=3.0.1=py37haa95532_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd3eb1b0_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=3.3.0=pyh44b312d_0
blinker=1.4=py37haa95532_0
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2021.5.25=haa95532_1
cached-property=1.5.2=py_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py37haa95532_0
cffi=1.14.5=py37hcd4344a_0
chardet=3.0.4=py37haa95532_1003
click=8.0.1=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
coverage=5.5=py37h2bbff1b_2
cryptography=3.4.7=py37h71e12ea_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py_2
cython=0.29.23=py37hd77b12b_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
freetype=2.10.4=h546665d_1
gast=0.2.2=py37_0
google-auth=1.32.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=py_0
grpcio=1.36.1=py37hc60d5dd_1
h5py=3.2.1=py37h3de5c98_0
hdf5=1.10.6=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.1=h0e60522_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37haa95532_0
intel-openmp=2021.2.0=haa95532_616
ipykernel=5.5.5=py37h7813e69_0
ipython=7.25.0=py37h7813e69_0
ipython_genutils=0.2.0=py_1
jbig=2.1=h8d14728_2003
jedi=0.18.0=py37h03978a9_2
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h8ffe710_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py37h03978a9_0
jupyter_server=1.9.0=pyhd8ed1ab_0
jupyterlab=3.0.16=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.6.0=pyhd8ed1ab_0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py37_0
keras-gpu=2.3.1=0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py37h8c56517_1
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libclang=11.1.0=default_h5c34c98_1
libdeflate=1.7=h8ffe710_5
libpng=1.6.37=h1d00b33_2
libprotobuf=3.14.0=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libtiff=4.3.0=h0c97f57_1
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=py37haa95532_0
markupsafe=2.0.1=py37hcc03f2d_0
matplotlib=3.4.2=py37h03978a9_0
matplotlib-base=3.4.2=py37h0d1fb12_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py37hcc03f2d_1003
mkl=2021.2.0=haa95532_296
mkl-service=2.3.0=py37h2bbff1b_1
mkl_fft=1.3.0=py37h277e83a_2
mkl_random=1.2.1=py37hf11a4ad_2
msys2-conda-epoch=20160418=1
multidict=5.1.0=py37h2bbff1b_2
nbclassic=0.3.1=pyhd8ed1ab_1
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.1.0=py37h03978a9_0
nbformat=5.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.1=pyhd8ed1ab_0
notebook=6.4.0=pyha770c72_0
numpy=1.20.2=py37ha4e8547_0
numpy-base=1.20.2=py37hc2deb75_0
oauthlib=3.1.0=py_0
olefile=0.46=pyh9f0ad1d_1
onnx=1.7.0=py37_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1k=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=20.9=pyh44b312d_0
pandoc=2.14.0.3=h8ffe710_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=8.2.0=py37h96663a1_1
pip=21.1.2=py37haa95532_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.19=pyha770c72_0
protobuf=3.14.0=py37hd77b12b_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd8ed1ab_0
pyjwt=1.7.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h03978a9_7
pyqt-impl=5.12.3=py37hf2a7229_7
pyqt5-sip=4.19.18=py37hf2a7229_7
pyqtchart=5.12=py37hf2a7229_7
pyqtwebengine=5.12.1=py37hf2a7229_7
pyreadline=2.1=py37_1
pyrsistent=0.17.3=py37hcc03f2d_2
pysocks=1.7.1=py37_1
python=3.7.10=h6244533_0
python-dateutil=2.8.1=py_0
python_abi=3.7=2_cp37m
pytz=2021.1=pyhd8ed1ab_0
pywin32=300=py37hcc03f2d_0
pywinpty=0.5.7=py37_0
pyyaml=5.4.1=py37h2bbff1b_1
pyzmq=22.1.0=py37hcce574b_0
qt=5.12.9=h5909a2a_4
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
requests-unixsocket=0.2.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.2=py37hf11a4ad_1
scipy=1.6.2=py37h66253e8_1
send2trash=1.7.1=pyhd8ed1ab_0
setuptools=52.0.0=py37haa95532_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=py37h03978a9_1
sqlite=3.36.0=h2bbff1b_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-addons=0.9.1=py37_1
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37haa95532_1
terminado=0.9.4=py37haa95532_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h8ffe710_1
tornado=6.1=py37hcc03f2d_1
traitlets=5.0.5=py_0
typeguard=2.7.0=py37_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.57.0=py37h03978a9_4
werkzeug=0.16.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yarl=1.6.3=py37h2bbff1b_0
zeromq=4.3.4=h0e60522_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.5.0=h6255e5f_0
